Delete Picture if exist, Bin File upload,folder management and all of table conditional update and insertion Yo

+++ server/controllers/upload.controller.js
@@ -2,16 +2,30 @@ import pool from '../config/connect-db.js';
 import imagekit from '../config/image-kit.js'
 
 const uploadFile = async (req, res) => {
-
+  const carId = req.body.car_id
   try {
     if (!req.files || req.files.length === 0) {
       return res.status(400).json({ error: "No files uploaded" });
     }
 
-    const [images] = await pool.query('select * from img_store where car_id = ?', [req.car_id])
+    const [images] = await pool.query('select * from img_store where car_id = ?', [carId])
+    const [binFile] = await pool.query('select * from file_store where car_id = ?', [carId])
 
-    for (const image of images) {
-      await imagekit.deleteFile(image.file_id)
+    if (images.length !== 0) {
+      for (const image of images) {
+        if (image && image.file_id) {
+          await imagekit.deleteFile(image.file_id)
+          await pool.query('delete from img_store where car_id = ?', [carId])
+        }
+      }
+    }
+    if (binFile.length !== 0) {
+      for (const bin of binFile) {
+        if (bin && bin.file_id) {
+          await imagekit.deleteFile(bin.file_id)
+          await pool.query('delete from file_store where car_id = ?', [carId])
+        }
+      }
     }
 
     const uploadedFiles = [];
@@ -19,14 +33,20 @@ const uploadFile = async (req, res) => {
     for (const file of req.files) {
       const fileExt = file.originalname.split(".").pop();
       const fileName = file.originalname.split(".")[0] + '-' + Math.floor(Math.random() * 4000 + 1000) + "." + fileExt;
+      const fileType = file.mimetype.split("/")[0]
 
       const uploaded = await imagekit.upload({
         file: file.buffer,
         fileName: fileName,
-        folder: "/jk-modules/images"
+        folder: `/jk-modules/${fileType === "image" ? 'images' : 'bins'}`
       });
 
-      await pool.query('insert into img_store (car_id,file_id,file_url) values (?,?,?)', [req.body.car_id, uploaded.fileId, uploaded.url])
+      if (fileType === "image") {
+        await pool.query('insert into img_store (car_id,file_id,file_url) values (?,?,?)', [carId, uploaded.fileId, uploaded.url])
+      }
+      else {
+        await pool.query('insert into file_store (file_id,car_id,file_url,file_name,archive_size) values (?,?,?,?,?)', [uploaded.fileId, carId, uploaded.url, fileName, uploaded.size])
+      }
 
       uploadedFiles.push({ file: uploaded.fileId, url: uploaded.url });
     }
